requirements.txt
flask
cairosvg
requests
bs4
pillow